+++ src/main/java/com/swyp8team2/vote/application/VoteService.java
@@ -3,8 +3,10 @@ package com.swyp8team2.vote.application;
 import com.swyp8team2.common.exception.BadRequestException;
 import com.swyp8team2.common.exception.ErrorCode;
 import com.swyp8team2.post.domain.Post;
+import com.swyp8team2.post.domain.PostImage;
 import com.swyp8team2.post.domain.PostRepository;
 import com.swyp8team2.post.domain.VoteType;
+import com.swyp8team2.vote.presentation.dto.PostImageVoteStatusResponse;
 import com.swyp8team2.user.domain.User;
 import com.swyp8team2.user.domain.UserRepository;
 import com.swyp8team2.vote.domain.Vote;
@@ -13,6 +15,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -23,6 +26,7 @@ public class VoteService {
     private final VoteRepository voteRepository;
     private final UserRepository userRepository;
     private final PostRepository postRepository;
+    private final RatioCalculator ratioCalculator;
 
     @Transactional
     public Long vote(Long voterId, Long postId, Long imageId) {
@@ -70,5 +74,33 @@ public class VoteService {
         Post post = postRepository.findById(vote.getPostId())
                 .orElseThrow(() -> new BadRequestException(ErrorCode.POST_NOT_FOUND));
         post.cancelVote(vote.getPostImageId());
+    }
+
+    public List<PostImageVoteStatusResponse> findVoteStatus(Long userId, Long postId) {
+        Post post = postRepository.findByIdFetchPostImage(postId)
+                .orElseThrow(() -> new BadRequestException(ErrorCode.POST_NOT_FOUND));
+        validateVoteStatus(userId, post);
+        int totalVoteCount = getTotalVoteCount(post.getImages());
+        return post.getImages().stream()
+                .map(image -> {
+                    String ratio = ratioCalculator.calculate(totalVoteCount, image.getVoteCount());
+                    return new PostImageVoteStatusResponse(image.getId(), image.getName(), image.getVoteCount(), ratio);
+                }).toList();
+    }
+
+    private void validateVoteStatus(Long userId, Post post) {
+        boolean voted = voteRepository.findByUserIdAndPostId(userId, post.getId())
+                .isPresent();
+        if (!(post.isAuthor(userId) || voted)) {
+            throw new BadRequestException(ErrorCode.ACCESS_DENIED_VOTE_STATUS);
+        }
+    }
+
+    private int getTotalVoteCount(List<PostImage> images) {
+        int totalVoteCount = 0;
+        for (PostImage image : images) {
+            totalVoteCount += image.getVoteCount();
+        }
+        return totalVoteCount;
     }
 }
