+++ src/main/java/com/swyp8team2/vote/presentation/VoteController.java
@@ -1,22 +1,21 @@
 package com.swyp8team2.vote.presentation;
 
 import com.swyp8team2.auth.domain.UserInfo;
-import com.swyp8team2.common.presentation.CustomHeader;
+import com.swyp8team2.vote.presentation.dto.PostImageVoteStatusResponse;
 import com.swyp8team2.vote.application.VoteService;
-import com.swyp8team2.vote.presentation.dto.ChangeVoteRequest;
 import com.swyp8team2.vote.presentation.dto.VoteRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -41,5 +40,13 @@ public class VoteController {
     ) {
         voteService.cancelVote(userInfo.userId(), voteId);
         return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/posts/{postId}/votes/status")
+    public ResponseEntity<List<PostImageVoteStatusResponse>> findVoteStatus(
+            @PathVariable("postId") Long postId,
+            @AuthenticationPrincipal UserInfo userInfo
+    ) {
+        return ResponseEntity.ok(voteService.findVoteStatus(userInfo.userId(), postId));
     }
 }
