+++ src/test/java/com/swyp8team2/post/presentation/PostControllerTest.java
@@ -197,38 +197,6 @@ class PostControllerTest extends RestDocsTest {
 
     @Test
     @WithMockUserInfo
-    @DisplayName("게시글 투표 상태 조회")
-    void findVoteStatus() throws Exception {
-        //given
-        var response = List.of(
-                new PostImageVoteStatusResponse(1L, "뽀또A", 2, "66.7"),
-                new PostImageVoteStatusResponse(2L, "뽀또B", 1, "33.3")
-        );
-        given(postService.findVoteStatus(1L, 1L))
-                .willReturn(response);
-
-        //when then
-        mockMvc.perform(RestDocumentationRequestBuilders.get("/posts/{postId}/status", 1)
-                        .header(HttpHeaders.AUTHORIZATION, "Bearer token"))
-                .andExpect(status().isOk())
-                .andExpect(content().json(objectMapper.writeValueAsString(response)))
-                .andDo(restDocs.document(
-                        requestHeaders(authorizationHeader()),
-                        pathParameters(
-                                parameterWithName("postId").description("게시글 Id")
-                        ),
-                        responseFields(
-                                fieldWithPath("[]").type(JsonFieldType.ARRAY).description("투표 선택지 목록"),
-                                fieldWithPath("[].id").type(JsonFieldType.NUMBER).description("이미지 Id"),
-                                fieldWithPath("[].imageName").type(JsonFieldType.STRING).description("사진 이름"),
-                                fieldWithPath("[].voteCount").type(JsonFieldType.NUMBER).description("투표 수"),
-                                fieldWithPath("[].voteRatio").type(JsonFieldType.STRING).description("투표 비율")
-                        )
-                ));
-    }
-
-    @Test
-    @WithMockUserInfo
     @DisplayName("게시글 삭제")
     void deletePost() throws Exception {
         //given
