+++ src/test/java/com/swyp8team2/vote/presentation/VoteControllerTest.java
@@ -1,18 +1,20 @@
 package com.swyp8team2.vote.presentation;
 
-import com.swyp8team2.common.presentation.CustomHeader;
+import com.swyp8team2.vote.presentation.dto.PostImageVoteStatusResponse;
 import com.swyp8team2.support.RestDocsTest;
 import com.swyp8team2.support.WithMockUserInfo;
-import com.swyp8team2.user.domain.Role;
-import com.swyp8team2.vote.presentation.dto.ChangeVoteRequest;
 import com.swyp8team2.vote.presentation.dto.VoteRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
+import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.restdocs.payload.JsonFieldType;
 
+import java.util.List;
+
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
@@ -21,8 +23,10 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
+import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 class VoteControllerTest extends RestDocsTest {
@@ -71,5 +75,37 @@ class VoteControllerTest extends RestDocsTest {
                         )
                 ));
         verify(voteService, times(1)).cancelVote(any(), any());
+    }
+
+    @Test
+    @WithMockUserInfo
+    @DisplayName("게시글 투표 상태 조회")
+    void findVoteStatus() throws Exception {
+        //given
+        var response = List.of(
+                new PostImageVoteStatusResponse(1L, "뽀또A", 2, "66.7"),
+                new PostImageVoteStatusResponse(2L, "뽀또B", 1, "33.3")
+        );
+        given(voteService.findVoteStatus(1L, 1L))
+                .willReturn(response);
+
+        //when then
+        mockMvc.perform(RestDocumentationRequestBuilders.get("/posts/{postId}/votes/status", 1)
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer token"))
+                .andExpect(status().isOk())
+                .andExpect(content().json(objectMapper.writeValueAsString(response)))
+                .andDo(restDocs.document(
+                        requestHeaders(authorizationHeader()),
+                        pathParameters(
+                                parameterWithName("postId").description("게시글 Id")
+                        ),
+                        responseFields(
+                                fieldWithPath("[]").type(JsonFieldType.ARRAY).description("투표 선택지 목록"),
+                                fieldWithPath("[].id").type(JsonFieldType.NUMBER).description("이미지 Id"),
+                                fieldWithPath("[].imageName").type(JsonFieldType.STRING).description("사진 이름"),
+                                fieldWithPath("[].voteCount").type(JsonFieldType.NUMBER).description("투표 수"),
+                                fieldWithPath("[].voteRatio").type(JsonFieldType.STRING).description("투표 비율")
+                        )
+                ));
     }
 }
